refactor: update widgets with eslint rules
