+++ src/widgets/news/index.ts
@@ -1,4 +1,1 @@
-import NewsFilters from "./ui/NewsFilters/NewsFilters";
-import NewsList from "./ui/NewsList/NewsList";
-
-export { NewsList, NewsFilters };
+export { NewsFilters, NewsList } from "./ui";
