+++ src/widgets/news/ui/NewsFilters/NewsFilters.tsx
@@ -1,19 +1,18 @@
-import { useAppDispatch } from "@/app/appStore";
-import { useTheme } from "@/app/providers/ThemeProvider";
-import Categories from "@/features/category/ui/Categories/Categories";
-import Search from "@/features/search/ui/Search/Search";
-import Slider from "@/features/slider/ui/Slider/Slider";
-import { IFilters } from "@/shared/interfaces";
-import { setFilters } from "@/entities/news/model/newsSlice";
+import { categoryModel } from "@/entities/category";
+import { newsModel } from "@/entities/news";
+import { Categories } from "@/features/category";
+import { Search } from "@/features/search";
+import { Slider } from "@/features/slider";
+import { useAppDispatch } from "@/shared/lib/redux";
+import { useTheme } from "@/shared/ui/Theme";
 import styles from "./styles.module.css";
-import { CategoriesType } from "@/entities/category";
 
 interface Props {
-  filters: IFilters;
-  categories: CategoriesType[];
+  filters: newsModel.IFilters;
+  categories: categoryModel.CategoriesType[];
 }
 
-const NewsFilters = ({ filters, categories }: Props) => {
+export const NewsFilters = ({ filters, categories }: Props) => {
   const { isDark } = useTheme();
 
   const dispatch = useAppDispatch();
@@ -26,7 +25,9 @@ const NewsFilters = ({ filters, categories }: Props) => {
             categories={categories}
             selectedCategory={filters.category}
             setSelectedCategory={(category) =>
-              dispatch(setFilters({ key: "category", value: category }))
+              dispatch(
+                newsModel.setFilters({ key: "category", value: category })
+              )
             }
           />
         </Slider>
@@ -35,11 +36,9 @@ const NewsFilters = ({ filters, categories }: Props) => {
       <Search
         keywords={filters.keywords}
         setKeywords={(keywords) =>
-          dispatch(setFilters({ key: "keywords", value: keywords }))
+          dispatch(newsModel.setFilters({ key: "keywords", value: keywords }))
         }
       />
     </div>
   );
 };
-
-export default NewsFilters;
