+++ src/widgets/news/ui/NewsFilters/index.ts
@@ -1,0 +1,1 @@
+export { NewsFilters } from "./NewsFilters";
