+++ src/widgets/news/ui/NewsList/index.ts
@@ -1,0 +1,1 @@
+export { NewsList } from "./NewsList";
