+++ src/widgets/news/ui/index.ts
@@ -1,0 +1,2 @@
+export { NewsFilters } from "./NewsFilters";
+export { NewsList } from "./NewsList";
